refactor/#252: 프로젝트 서비스 레이어 기능별로 분리 (Command/Query)
