+++ src/main/java/inha/git/project/api/controller/ProjectRecommendController.java
@@ -2,7 +2,7 @@ package inha.git.project.api.controller;
 
 import inha.git.common.BaseResponse;
 import inha.git.project.api.controller.dto.request.RecommendRequest;
-import inha.git.project.api.service.ProjectRecommendService;
+import inha.git.project.api.service.recommend.ProjectRecommendService;
 import inha.git.user.domain.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
